Moving classes around a bit for better file structure
